backend/requirements.txt
fastapi==0.111.0
uvicorn>=0.29.0
greenlet>=3.1.0
python-multipart==0.0.9
pandas>=2.1.0
openpyxl>=3.1.0
Pillow>=10.0.0
python-bidi==0.4.2
arabic-reshaper==3.0.0
playwright>=1.47.0
websockets==12.0
aiofiles==23.2.1
pydantic>=2.9.0
phonenumbers==8.13.36
